+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransactionFactory.java
@@ -36,9 +36,22 @@ public class TransactionFactory {
         return RegisterDataTransaction.builder();
     }
 
+    /**
+     * Creates a new {@link TransferScheduleTransaction.TransferScheduleTransactionBuilder} for
+     * creating a {@link TransferScheduleTransaction}
+     *
+     * @return the builder for a {@link TransferScheduleTransaction}
+     */
     public static TransferScheduleTransaction.TransferScheduleTransactionBuilder newScheduleData() {
         return TransferScheduleTransaction.builder();
     }
+
+    /**
+     * Creates a new {@link TransferScheduleWithMemoTransaction.TransferScheduleWithMemoTransactionBuilder} for
+     * creating a {@link TransferScheduleWithMemoTransaction}
+     *
+     * @return the builder for a {@link TransferScheduleWithMemoTransaction}
+     */
     public static TransferScheduleWithMemoTransaction.TransferScheduleWithMemoTransactionBuilder newScheduleDataWithMemo() {
         return TransferScheduleWithMemoTransaction.builder();
     }