+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransferScheduleWithMemo.java
@@ -8,6 +8,9 @@ import lombok.val;
 
 import java.nio.ByteBuffer;
 
+/**
+ * Transfer an amount with schedule and memo.
+ */
 @ToString
 @Getter
 @EqualsAndHashCode(callSuper = true)